requirements.txt
gql>=2.0.0
pytest>=6.2.2
pytest_bdd>=4.0.2
PyYAML>=5.4.1
nested_lookup
xmltodict>=0.12.0
google-cloud-bigquery
google-api-core
requests>=2.25.1
loguru>=0.5.3
pytest_check
Shapely>=1.7.1